move calc_bivariate_qq&pdf from cli.py to utils.py
