+++ precimed/mixer_dev/cli.py
@@ -46,8 +46,8 @@ from .utils_obsolete import UnivariateParams_obsolete
 
 from bivar_mixer.utils import calc_qq_plot
 from bivar_mixer.utils import calc_power_curve
-from bivar_mixer.cli import calc_bivariate_qq
-from bivar_mixer.cli import calc_bivariate_pdf
+from bivar_mixer.utils import calc_bivariate_qq
+from bivar_mixer.utils import calc_bivariate_pdf
 
 import common.utils_cli
